🚚 zv: Rename `dict_utils` mod to `as_value`
This is a more appropriate name and we'll move SerializeValue and
DeserializeValue into this mod too now.
