+++ book/src/faq.md
@@ -18,7 +18,7 @@ use zbus::{
     proxy, interface, fdo::Result,
     zvariant::{
         Type,
-        dict_utils::{opt_value, value},
+        as_value::{opt_value, value},
     },
 };
 use serde::{Deserialize, Serialize};
@@ -63,14 +63,14 @@ impl DictionaryGiverInterface {
 ```
 
 Since the fields have to be transformed from/into `zvariant::Value`, make sure to use the `with`
-attribute with the appropriate helper module from `zvariant::dict_utils` module.
+attribute with the appropriate helper module from `zvariant::as_value` module.
 
 Moroever, since D-Bus does not have a concept of nullable types, it's important to ensure that
 `skip_serializing_if` and `default` attributes are used for optional fields. Fortunately, you can
 make use of the `default` container attribute if your struct can implemented `Default` trait:
 
 ```rust,noplayground
-use zbus::zvariant::{Type, dict_utils::{opt_value, value}};
+use zbus::zvariant::{Type, as_value::{opt_value, value}};
 # use serde::{Deserialize, Serialize};
 
 #[derive(Default, Deserialize, Serialize, Type)]
@@ -92,7 +92,7 @@ struct:
 
 ```rust,noplayground
 use std::collections::HashMap;
-use zbus::zvariant::{Type, OwnedValue, dict_utils::{value}};
+use zbus::zvariant::{Type, OwnedValue, as_value::{value}};
 # use serde::{Deserialize, Serialize};
 
 #[derive(Default, Deserialize, Serialize, Type)]
